requirements.txt
phidata
python-dotenv
groq
duckduckgo-search
googlesearch-python
pycountry
packaging
openai
wikipedia
newspaper4k 
lxml_html_clean
fastapi[standard]
sqlalchemy
yfinance
duckdb
streamlit
watchdog
pandas
scikit-learn